requirements.txt
bcrypt==4.0.1
passlib==1.7.4
fastapi==0.115.0
uvicorn[standard]==0.31.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic[email]==2.9.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
